# apps/client/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SCTG Vite React Heroui Auth0 Template API",
    "version": "1.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Root endpoint",
        "description": "Returns a simple greeting.",
        "responses": {
          "200": {
            "description": "A simple greeting message.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Hello World!"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "description": "Verify if the worker is up and running.",
        "responses": {
          "200": {
            "description": "Worker is healthy.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/__auth0/token": {
      "post": {
        "summary": "Get Auth0 Management Token (auth0:admin:api)",
        "description": "Requests an Auth0 Management API token.\n\n**Required Permission**: `auth0:admin:api`\n",
        "security": [
          {
            "bearerAuth": [
              "auth0:admin:api"
            ]
          },
          {
            "oauth2": [
              "auth0:admin:api"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Auth0Token"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden."
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/__auth0/autopermissions": {
      "post": {
        "summary": "Auto-assign permissions (JWT required)",
        "description": "Assigns default permissions to the current user if missing.\n\n**Requirements**: A valid JWT (authentication required, but no specific scope needed).\n",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "oauth2": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutoPermissionsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/ping": {
      "get": {
        "summary": "Protected ping (read:api)",
        "description": "Verify authentication and basic connectivity.\n\n**Required Permission**: `read:api`\n",
        "security": [
          {
            "bearerAuth": [
              "read:api"
            ]
          },
          {
            "oauth2": [
              "read:api"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PingResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden."
          }
        }
      }
    },
    "/api/get_user": {
      "get": {
        "summary": "Get user ID (read:api)",
        "description": "Returns the sub claim of the authenticated user.\n\n**Required Permission**: `read:api`\n",
        "security": [
          {
            "bearerAuth": [
              "read:api"
            ]
          },
          {
            "oauth2": [
              "read:api"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/get/{user}": {
      "get": {
        "summary": "Debug request info (read:api)",
        "description": "Returns detailed session and request info.\n\n**Required Permission**: `read:api`\n",
        "security": [
          {
            "bearerAuth": [
              "read:api"
            ]
          },
          {
            "oauth2": [
              "read:api"
            ]
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Dynamic user parameter."
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DebugResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "oauth2": {
        "type": "oauth2",
        "description": "Auth0 OAuth2 Authorization Code Flow",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://sctg-development.eu.auth0.com/authorize",
            "tokenUrl": "https://sctg-development.eu.auth0.com/oauth/token",
            "scopes": {
              "openid": "OpenID Connect",
              "profile": "Profile",
              "email": "Email",
              "read:api": "Read API",
              "write:api": "Write API",
              "auth0:admin:api": "Auth0 Admin API"
            }
          }
        }
      }
    },
    "schemas": {
      "Auth0Token": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "The Auth0 Management API access token."
          },
          "token_type": {
            "type": "string",
            "example": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "description": "Token lifetime in seconds."
          },
          "from_cache": {
            "type": "boolean",
            "description": "Indicates if the token was retrieved from KV cache."
          }
        }
      },
      "AutoPermissionsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "added": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of permissions successfully added to the user."
          },
          "message": {
            "type": "string",
            "description": "Optional status message."
          }
        }
      },
      "PingResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "user": {
            "type": "string",
            "nullable": true,
            "description": "The user's sub claim."
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "user": {
            "type": "string",
            "description": "The user's sub ID."
          }
        }
      },
      "DebugResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "user": {
            "type": "string"
          },
          "sub": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "token": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "oauth2": []
    }
  ]
}